Partial View

+++ MyLeassing.Web/Controllers/PropertyTypesController.cs
@@ -125,26 +125,30 @@ namespace MyLeassing.Web.Controllers
             }
 
             var propertyType = await _context.PropertyTypes
+                .Include(pt =>pt.Properties)
                 .FirstOrDefaultAsync(m => m.Id == id);
             if (propertyType == null)
             {
                 return NotFound();
             }
-
-            return View(propertyType);
+            if(propertyType.Properties.Count !=0)
+            {
+                ModelState.AddModelError(string.Empty, "The record could not be deleted, because it contains properties.");
+                return RedirectToAction(nameof(Index));
+            }
+            try
+            {
+                _context.PropertyTypes.Remove(propertyType);
+                await _context.SaveChangesAsync();
+                return RedirectToAction(nameof(Index));
+            }
+            catch(Exception ex)
+            {
+                ModelState.AddModelError(string.Empty, ex.ToString());
+                return RedirectToAction(nameof(Index));
+            }
         }
-
-        // POST: PropertyTypes/Delete/5
-        [HttpPost, ActionName("Delete")]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed(int id)
-        {
-            var propertyType = await _context.PropertyTypes.FindAsync(id);
-            _context.PropertyTypes.Remove(propertyType);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
-        }
-
+                
         private bool PropertyTypeExists(int id)
         {
             return _context.PropertyTypes.Any(e => e.Id == id);
